+++ packages/manager/src/features/Linodes/LinodeCreatev2/Tabs/Backups/LinodeSelect.test.tsx
@@ -1,16 +1,8 @@
 import React from 'react';
 
-import { linodeFactory } from 'src/factories';
-import { makeResourcePage } from 'src/mocks/serverHandlers';
-import { HttpResponse, http, server } from 'src/mocks/testServer';
-import {
-  mockMatchMedia,
-  renderWithThemeAndHookFormContext,
-} from 'src/utilities/testHelpers';
+import { renderWithThemeAndHookFormContext } from 'src/utilities/testHelpers';
 
 import { LinodeSelect } from './LinodeSelect';
-
-beforeAll(() => mockMatchMedia());
 
 describe('LinodeSelect', () => {
   it('should render a heading', () => {
@@ -22,49 +14,5 @@ describe('LinodeSelect', () => {
 
     expect(heading).toBeVisible();
     expect(heading.tagName).toBe('H2');
-  });
-
-  it('should render Linodes from the API', async () => {
-    const linodes = linodeFactory.buildList(10);
-
-    server.use(
-      http.get('*/linode/instances*', () => {
-        return HttpResponse.json(makeResourcePage(linodes));
-      })
-    );
-
-    const { findByText } = renderWithThemeAndHookFormContext({
-      component: <LinodeSelect />,
-    });
-
-    for (const linode of linodes) {
-      // eslint-disable-next-line no-await-in-loop
-      await findByText(linode.label);
-    }
-  });
-
-  it('should select a linode based on form state', async () => {
-    const selectedLinode = linodeFactory.build({
-      id: 1,
-      label: 'my-selected-linode',
-    });
-
-    server.use(
-      http.get('*/linode/instances*', () => {
-        return HttpResponse.json(makeResourcePage([selectedLinode]));
-      })
-    );
-
-    const { findByLabelText } = renderWithThemeAndHookFormContext({
-      component: <LinodeSelect />,
-      useFormOptions: {
-        defaultValues: { linode: selectedLinode },
-      },
-    });
-
-    const radio = await findByLabelText(selectedLinode.label);
-
-    expect(radio).toBeEnabled();
-    expect(radio).toBeChecked();
   });
 });
